+++ R/univariate_1freq_plot.R
@@ -1,3 +1,7 @@
+#' @importFrom ggplot2 autoplot
+#' @export
+ggplot2::autoplot
+
 #' Automatically create a ggplot for objects obtained from univariate_frequency()
 #'
 #' @description Takes an object produced by \code{univariate_frequency()}, and plots the claim frequency.
@@ -6,9 +10,11 @@
 #' @param background show exposure as a histogram (default is TRUE)
 #' @param labels show labels with the exposure (default is TRUE)
 #' @param sort sort (or order) risk factor into descending order by exposure (default is FALSE)
+#' @param sort_manual sort (or order) risk factor into own ordering; should be a character vector (default is NULL)
 #' @param dec.mark control the format of the decimal point, as well as the mark between intervals before the decimal point, choose either "," (default) or "."
 #' @param color change the color of the points and line ("dodgerblue" is default)
 #' @param color_bg change the color of the histogram ("#E7B800" is default)
+#' @param label_width width of labels on the x-axis (10 is default)
 #'
 #' @import ggplot2
 #' @importFrom stringr str_wrap
@@ -21,7 +27,7 @@
 #' autoplot(x, sort = TRUE, dec.mark = ".", color_bg = "mediumseagreen")
 #'
 #' @export
-autoplot.univ_freq <- function(x, background = TRUE, labels = TRUE, sort = FALSE, dec.mark = ",", color = "dodgerblue", color_bg = "#E7B800"){
+autoplot.univ_freq <- function(x, background = TRUE, labels = TRUE, sort = FALSE, sort_manual = NULL, dec.mark = ",", color = "dodgerblue", color_bg = "#E7B800", label_width = 10){
 
   df <- x$df
   xvar <- x$xvar
@@ -49,9 +55,9 @@ autoplot.univ_freq <- function(x, background = TRUE, labels = TRUE, sort = FALSE
     df[[exposure]] <- round(df[[exposure]], 0)
 
     hist_bg <- list(
-      geom_bar(data = df, aes(x = .data[[xvar]], y = .data[["exposure_scale"]]),
+      ggplot2::geom_bar(data = df, aes(x = .data[[xvar]], y = .data[["exposure_scale"]]),
                      stat = "identity", color = color_bg, fill = color_bg, alpha = 0.4),
-      scale_y_continuous(sec.axis = sec_axis(~ . * max(df[[exposure]]) / max(df$frequency),
+      ggplot2::scale_y_continuous(sec.axis = sec_axis(~ . * max(df[[exposure]]) / max(df$frequency),
                                              name = exposure,
                                              labels = sep_fn),
                          labels = sep_fn,
@@ -63,20 +69,29 @@ autoplot.univ_freq <- function(x, background = TRUE, labels = TRUE, sort = FALSE
 
   if (isTRUE( background ) & isTRUE( labels )) {
     labels_bg <- list(
-      geom_text(data = df, aes(x = .data[[xvar]], y = .data[["exposure_scale"]], label = sep_fn(.data[[exposure]])),
+      ggplot2::geom_text(data = df, aes(x = .data[[xvar]], y = .data[["exposure_scale"]], label = sep_fn(.data[[exposure]])),
                 vjust = "inward", size = 3)
     )
   }
 
-  p1 <- ggplot(data = df) +
+  if ( !is.null( sort_manual )){
+    hist_sort <- list(
+      ggplot2::scale_x_discrete(labels = function(x) stringr::str_wrap(x, width = label_width), limits = sort_manual )
+    )
+  } else {
+    hist_sort <- list(
+      ggplot2::scale_x_discrete(labels = function(x) stringr::str_wrap(x, width = label_width) )
+    )}
+
+  p1 <- ggplot2::ggplot(data = df) +
     {if( isTRUE( background )) hist_bg} +
-    geom_point(aes(x = .data[[xvar]], y = .data[["frequency"]]), color = color) +
-    geom_line(aes(x = .data[[xvar]], y = .data[["frequency"]], group = 1), color = color) +
-    {if( isTRUE( background ) & isTRUE( labels )) labels_bg} +
-    theme_minimal() +
-    labs(y = "Frequency", x = xvar) +
-    {if( !isTRUE ( background )) scale_y_continuous(labels = sep_fn)} +
-    scale_x_discrete(labels = function(x) stringr::str_wrap(x, width = 10))
+    ggplot2::geom_point(aes(x = .data[[xvar]], y = .data[["frequency"]]), color = color) +
+    ggplot2::geom_line(aes(x = .data[[xvar]], y = .data[["frequency"]], group = 1), color = color) +
+    { if( isTRUE( background ) & isTRUE( labels )) labels_bg } +
+    ggplot2::theme_minimal() +
+    ggplot2::labs(y = "Frequency", x = xvar) +
+    { if( !isTRUE ( background )) ggplot2::scale_y_continuous(labels = sep_fn) } +
+    { if( !is.null( sort_manual )) hist_sort }
 
   return(p1)
 }